requirements.txt
dash==2.8.1
dash_bootstrap_components==1.3.1
dash_core_components==2.0.0
numpy==1.24.1
opencv_python==4.7.0.72
Pillow==9.5.0
plotly==5.13.0
PyAutoGUI==0.9.53
scipy==1.10.1
